+++ lib/data/models/search_input_model.dart
