+++ lib/data/models/task_model.dart
@@ -1,0 +1,21 @@
+import 'package:hive/hive.dart';
+
+part 'task_model.g.dart';
+
+@HiveType(typeId: 0)
+class Task extends HiveObject {
+  @HiveField(0)
+  String title;
+
+  @HiveField(1)
+  String description;
+
+  @HiveField(2)
+  bool isCompleted;
+
+  Task({
+    required this.title,
+    required this.description,
+    required this.isCompleted,
+  });
+}
