--- lib/modals/task_modal.dart
@@ -1,11 +1,0 @@
-class Task {
-  String title;
-  String description;
-  bool isCompleted;
-
-  Task({
-    required this.title,
-    required this.description,
-    this.isCompleted = false,
-  });
-}
