+++ lib/ui/widgets/search/search_input_widget.dart
@@ -1,0 +1,76 @@
+import 'package:flutter/material.dart';
+import 'package:taski_todo/core/constants/app_colors.dart';
+import 'package:taski_todo/core/constants/app_text_styles.dart';
+
+class SearchInput extends StatefulWidget {
+  final TextEditingController textController;
+  final String hintText;
+  final ValueChanged<String>? onChanged;
+
+  const SearchInput({
+    super.key,
+    required this.textController,
+    required this.hintText,
+    this.onChanged,
+  });
+
+  @override
+  _SearchInputState createState() => _SearchInputState();
+}
+
+class _SearchInputState extends State<SearchInput> {
+  @override
+  void initState() {
+    super.initState();
+    widget.textController.addListener(_onTextChanged);
+  }
+
+  @override
+  void dispose() {
+    widget.textController.removeListener(_onTextChanged);
+    super.dispose();
+  }
+
+  void _onTextChanged() => setState(() {});
+
+  @override
+  Widget build(BuildContext context) {
+    return TextField(
+      controller: widget.textController,
+      onChanged: widget.onChanged,
+      style: AppTextStyles.subtitlesearch,
+      decoration: InputDecoration(
+        hintText: widget.hintText,
+        hintStyle: AppTextStyles.subtitlesearch,
+        prefixIcon: const Icon(Icons.search, color: Colors.blue),
+        fillColor: AppColors.paleWhite50,
+        filled: true,
+        border: InputBorder.none,
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(16),
+          borderSide: const BorderSide(
+            color: Colors.blue,
+            width: 2.0,
+          ),
+        ),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(16),
+          borderSide: const BorderSide(
+            color: Colors.transparent,
+          ),
+        ),
+        suffixIcon: widget.textController.text.isNotEmpty
+            ? IconButton(
+                icon: const Icon(Icons.cancel_sharp,
+                    size: 16, color: AppColors.stateBlue),
+                onPressed: () {
+                  widget.textController.clear();
+                  widget.onChanged?.call('');
+                  setState(() {});
+                },
+              )
+            : null,
+      ),
+    );
+  }
+}
