+++ pubspec.yaml
@@ -23,6 +23,10 @@ dependencies:
   provider: ^6.1.2
   flutter_native_splash: ^2.4.4
   flutter_launcher_icons: ^0.14.2
+  
+  hive_flutter: ^1.1.0
+  hive_generator: ^2.0.1
+  build_runner: ^2.4.14
 
 
 flutter_native_splash:
